+++ sql-planner/src/executor/engine/helpers.rs
@@ -1262,6 +1262,16 @@ pub fn materialize_values(
         }
         vtable
     } else {
+        let limit = exec_plan.get_ir_plan().effective_options.sql_motion_row_max;
+        let values_count = children.len();
+        if limit > 0 && limit < values_count as i64 {
+            return Err(SbroadError::ExecutionError(format_smolstr!(
+                "Exceeded maximum number of rows ({}) in virtual table: {}",
+                limit,
+                values_count,
+            )));
+        }
+
         let first_row_id = children
             .first()
             .expect("Values node must contain children.");