+++ src/sql/dispatch.rs
@@ -1344,7 +1344,7 @@ fn port_write_tuples<'tuples, 'p>(
         })?;
         *row_count += 1;
         if max_rows > 0 && *row_count > max_rows {
-            return Err(SbroadError::DispatchError(format_smolstr!(
+            return Err(SbroadError::ExecutionError(format_smolstr!(
                 "Exceeded maximum number of rows ({max_rows}) in virtual table: {row_count}"
             )));
         }